feat: cambia precio a Gs sin decimales (INT)
- modelo Producto precio BigDecimal -> int (Gs)
- DAO mapRow getInt/setInt para precio
- Controlador validar int >0 y parse con limpieza de puntos de miles (1.500.000)
- VentanaPrincipal label Precio (Gs), columna Precio (Gs) tipo Integer, mostrar int
- VentanaPrincipal.form columna y label actualizados
- query.sql precio DECIMAL(10,2) -> INT, seed con valores Gs (7.500.000, 150.000, 450.000, 1.200.000)
- README actualiza validacion precio

+++ src/dao/ProductoDAO.java
@@ -3,7 +3,6 @@ package dao;
 import config.Conexion;
 import modelo.Producto;
 
-import java.math.BigDecimal;
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
@@ -18,7 +17,7 @@ public class ProductoDAO {
         p.setCodigo(rs.getString("codigo"));
         p.setNombre(rs.getString("nombre"));
         p.setCategoria(rs.getString("categoria"));
-        p.setPrecio(rs.getBigDecimal("precio"));
+        p.setPrecio(rs.getInt("precio"));
         p.setStock(rs.getInt("stock"));
         p.setEstado(rs.getString("estado"));
         return p;
@@ -110,7 +109,7 @@ public class ProductoDAO {
             ps.setString(1, p.getCodigo());
             ps.setString(2, p.getNombre());
             ps.setString(3, p.getCategoria());
-            ps.setBigDecimal(4, p.getPrecio());
+            ps.setInt(4, p.getPrecio());
             ps.setInt(5, p.getStock());
             ps.setString(6, p.getEstado());
             ps.executeUpdate();
@@ -127,7 +126,7 @@ public class ProductoDAO {
             ps.setString(1, p.getCodigo());
             ps.setString(2, p.getNombre());
             ps.setString(3, p.getCategoria());
-            ps.setBigDecimal(4, p.getPrecio());
+            ps.setInt(4, p.getPrecio());
             ps.setInt(5, p.getStock());
             ps.setString(6, p.getEstado());
             ps.setInt(7, p.getId());
