+++ src/modelo/Producto.java
@@ -1,19 +1,17 @@
 package modelo;
 
-import java.math.BigDecimal;
-
 public class Producto {
     private int id;
     private String codigo;
     private String nombre;
     private String categoria;
-    private BigDecimal precio;
+    private int precio; // Gs, sin decimales
     private int stock;
     private String estado;
 
     public Producto() {}
 
-    public Producto(int id, String codigo, String nombre, String categoria, BigDecimal precio, int stock, String estado) {
+    public Producto(int id, String codigo, String nombre, String categoria, int precio, int stock, String estado) {
         this.id = id;
         this.codigo = codigo;
         this.nombre = nombre;
@@ -23,7 +21,7 @@ public class Producto {
         this.estado = estado;
     }
 
-    public Producto(String codigo, String nombre, String categoria, BigDecimal precio, int stock, String estado) {
+    public Producto(String codigo, String nombre, String categoria, int precio, int stock, String estado) {
         this(0, codigo, nombre, categoria, precio, stock, estado);
     }
 
@@ -39,8 +37,8 @@ public class Producto {
     public String getCategoria() { return categoria; }
     public void setCategoria(String categoria) { this.categoria = categoria; }
 
-    public BigDecimal getPrecio() { return precio; }
-    public void setPrecio(BigDecimal precio) { this.precio = precio; }
+    public int getPrecio() { return precio; }
+    public void setPrecio(int precio) { this.precio = precio; }
 
     public int getStock() { return stock; }
     public void setStock(int stock) { this.stock = stock; }
